fix(rpc/v2/envs): move into correct package and fixup validation assertions
