+++ rpc/v2/environments/validation.go
@@ -1,4 +1,4 @@
-package configuration
+package environments
 
 import "go.flipt.io/flipt/errors"
 
@@ -19,7 +19,11 @@ func (r *GetResourceRequest) Validate() error {
 }
 
 func (r *ListResourcesRequest) Validate() error {
-	return requireType(r)
+	if err := requireType(r); err != nil {
+		return err
+	}
+
+	return requireNamespace(r)
 }
 
 func (r *UpdateResourceRequest) Validate() error {
@@ -31,6 +35,10 @@ func (r *DeleteResourceRequest) Validate() error {
 }
 
 func requireKey(k keyed) error {
+	if k.GetEnvironment() == "" {
+		return errors.ErrInvalid("environment must not be empty")
+	}
+
 	if k.GetKey() == "" {
 		return errors.ErrInvalid("key must not be empty")
 	}
@@ -39,12 +47,16 @@ func requireKey(k keyed) error {
 }
 
 func requireType(r typed) error {
-	if r.GetNamespace() == "" {
-		return errors.ErrInvalid("namespace must not be empty")
+	if r.GetTypeUrl() == "" {
+		return errors.ErrInvalid("type must not be empty")
 	}
 
-	if r.GetType() == "" {
-		return errors.ErrInvalid("type must not be empty")
+	return nil
+}
+
+func requireNamespace(r namespaced) error {
+	if r.GetNamespace() == "" {
+		return errors.ErrInvalid("namespace must not be empty")
 	}
 
 	return nil
@@ -52,9 +64,14 @@ func requireType(r typed) error {
 
 func requireResource(r interface {
 	keyed
+	namespaced
 	typed
 }) error {
 	if err := requireType(r); err != nil {
+		return err
+	}
+
+	if err := requireNamespace(r); err != nil {
 		return err
 	}
 
@@ -62,10 +79,14 @@ func requireResource(r interface {
 }
 
 type keyed interface {
+	GetEnvironment() string
 	GetKey() string
 }
 
 type typed interface {
-	GetType() string
+	GetTypeUrl() string
+}
+
+type namespaced interface {
 	GetNamespace() string
 }
